backend/requirements.txt
flask
sqlalchemy
psycopg2-binary
werkzeug
flask-cors
pyjwt
bcrypt
